requirements.txt
Django==2.2.19
djangorestframework==3.12.4
pytz==2022.1
sqlparse==0.4.2
djoser==2.1.0
Pillow==9.2.0 
django-colorfield==0.7.2
django-filter==21.1
drf-extra-fields==3.4.0
borb==2.0.32